+++ src/CAServer.ContractEventHandler/appsettings.json
@@ -46,8 +46,8 @@
         "PrivateKey": "***",
         "IsMainChain": true
       },
-      "tDVV": {
-        "ChainId": "tDVV",
+      "tDVW": {
+        "ChainId": "tDVW",
         "BaseUrl": "http://127.0.0.1:8000",
         "ContractAddress": "***",
         "TokenContractAddress": "***",
@@ -73,7 +73,7 @@
     }
   },
   "GraphQL": {
-    "GraphQLConnection": "http://127.0.0.1:8083/Indexer_DApp/PortKeyIndexerCASchema/graphql"
+    "Configuration": "http://127.0.0.1:8083/Indexer_DApp/PortKeyIndexerCASchema/graphql"
   },
   "Sync": {
     "Sync": 30,
